move Blog type
