+++ app/page.tsx
@@ -1,10 +1,9 @@
 import Image from "next/image";
 import styles from "./page.module.css";
 import { Header } from "@/components/header/header";
-import { BlogItem, BlogMenu } from "@/components/blog";
+import { BlogItem, BlogMenu, Blog } from "@/components/blog";
 import { Footer } from "@/components/footer/footer";
 import { getBlogs } from "@/api/rest";
-import { Blog } from "@/api/types";
 
 const Home = async () => {
     const blogs: Blog[] = await getBlogs();
