+++ components/blog/index.ts
@@ -1,2 +1,3 @@
 export * from './blog-item';
 export * from './blog-menu';
+export * from './types';
