+++ components/footer/footer.tsx
@@ -1,4 +1,3 @@
-import { Blog } from "@/api/types";
 import { FC } from "react";
 import styles from "./footer.module.css";
